python_tts/requirements.txt
boto3
contextlib2
azure.cognitiveservices.speech
pandas
openpyxl
logging
ushlex
subprocess.run